Adicionado métodos da api para escrita no banco de Fornecedores

+++ MinhaApiCompleta/src/DevIO.Business/Services/FornecedorService.cs
@@ -18,31 +18,33 @@ namespace DevIO.Business.Services
             _fornecedorRepository = fornecedorRepository;
             _enderecoRepository = enderecoRepository;
         }
-        public async Task Adicionar(Fornecedor fornecedor)
+        public async Task<bool> Adicionar(Fornecedor fornecedor)
         {
             if(!ExecutarValidacao(new FornecedorValidation(), fornecedor) 
-                || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
+                || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return false;
             
             if(_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any()) 
             {
                 Notificar("Já existe um fornecedor com o documento informado");
-                return;
+                return false;
             }
 
             await _fornecedorRepository.Adicionar(fornecedor);
+            return true;
         }
 
-        public async Task Atualizar(Fornecedor fornecedor)
+        public async Task<bool> Atualizar(Fornecedor fornecedor)
         {
-            if(!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
+            if(!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return false;
 
             if(_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any()) 
             {
                 Notificar("Já existe um fornecedor com o documento informado");
-                return;
+                return false;
             }
 
             await _fornecedorRepository.Atualizar(fornecedor);
+            return true;
         }
 
         public async Task AtualizarEndereco(Endereco endereco)
@@ -51,19 +53,24 @@ namespace DevIO.Business.Services
             await _enderecoRepository.Atualizar(endereco);
         }
 
-        public async Task Remover(Guid id)
+        public async Task<bool> Remover(Guid id)
         {
             if (_fornecedorRepository.ObterFornecedorProdutosEndereco(id).Result.Produtos.Any())
             {
                 Notificar("O fornecedor possui produtos cadastrados");
-                return;
+                return false;
             }
 
             var enderecoFornecedor = await _enderecoRepository.ObterEnderecoPorFornecedor(id);
 
-
-            await _enderecoRepository.Remover(enderecoFornecedor.Id);
+            if(enderecoFornecedor != null)
+            {
+                await _enderecoRepository.Remover(enderecoFornecedor.Id);
+            }
+            
             await _fornecedorRepository.Remover(id);
+
+            return true;
         }
 
         public void Dispose()
